requirements.txt
fastapi==0.121.1
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

qdrant-client==1.11.1
requests==2.32.4
numpy==2.0.1
prometheus-client==0.20.0
starlette==0.49.3

streamlit==1.51.0
pillow==12.1.1
pypdf==6.7.4
tqdm==4.66.5
